packages/backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
strawberry-graphql>=0.129
httpx>=0.24
python-dotenv>=1.0.0
pdfplumber>=0.7.5
python-multipart>=0.0.6
openai>=1.0.0